Define context as a pair root value and schema
Application Rule propagates a new context down to nested subterms. A
context is made of a root value and a schema.

Alternatively, we could redefine the Application Rule as

... and \Gamma', v' |- t -> t'
-------------------------------------
\Gamma, v |- m_(x:u).t -> m_(x:u).t'

where t is evaluated in the new context created by the Application Rule.
In this case, we would not need context as a term.

However we want to keep track of partial results in evaluation steps.
For example, we may want to return partial results to the client before
evaluating every field in t.