Change nest rule
The schema in the premise of the nest rule is required to be the
projection of the conclusion schema.

This may seem not strictly necessary since the application of the context rule
will override the schema anyway.

However, if the premise schema is bottom as before, we could apply also
the empty rule. Which is wrong.