Replace set with parallel construct
We want a single term m(x:t1).t2 to match a schema (... m(x:T1) ...).
So basically we need to say that m(x:t1).t2 is actually a set with a
single element.

Note that we use | for both selection sets and object types.

Finally, it seems to be easier to prove some properties of the language
when terms and types are represented in this way.